Deleted build from git, added corrections and services

+++ src/controllers/stateController.ts
@@ -1,91 +1,110 @@
 import type { Request, Response } from 'express'
 import { PrismaClient } from '@prisma/client'
+import { httpStatus } from '../utils/httpStatus'
+import { createError, successResponse } from '../utils/responseHandler'
 
 const prisma = new PrismaClient()
 
 export const getAllStates = async (req: Request, res: Response): Promise<Response> => {
-  const tenantId = res.locals.tenant_id
   try {
-    const states = await prisma.state.findMany({
-      where: {
-        tenant_id: tenantId
-      }
-    })
-    return res.json(states)
+    const states = await prisma.state.findMany()
+    return res.status(httpStatus.OK).json(successResponse(states, httpStatus.OK, 'States retrieved successfully'))
   } catch (error: any) {
-    return res.status(500).json({ error: error.message })
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).json(
+      createError(httpStatus.INTERNAL_SERVER_ERROR, error.message as string)
+    )
   }
 }
 
 export const getState = async (req: Request, res: Response): Promise<Response> => {
-  const tenantId = res.locals.tenant_id
   const stateId = parseInt(req.params.id)
   try {
     const state = await prisma.state.findUnique({
       where: {
-        tenant_id: tenantId,
         id: stateId
       }
     })
-    return (state != null) ? res.json(state) : res.status(404).send('State not found')
+    return (state != null)
+      ? res.status(httpStatus.OK).json(successResponse(state, httpStatus.OK, 'State retrieved successfully'))
+      : res.status(httpStatus.OK).json(successResponse(state, httpStatus.OK, 'State not found'))
   } catch (error: any) {
-    return res.status(500).json({ error: error.message })
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).json(
+      createError(httpStatus.INTERNAL_SERVER_ERROR, error.message as string)
+    )
   }
 }
 
 export const createState = async (req: Request, res: Response): Promise<Response> => {
-  const tenantId = res.locals.tenant_id
-
-  if (!tenantId) {
-    return res.status(400).send('Tenant ID is required')
-  }
-
   const { description } = req.body
   try {
     const newState = await prisma.state.create({
       data: {
-        tenant_id: tenantId,
         description
       }
     })
-    return res.json(newState)
+    return res.status(httpStatus.OK).json(successResponse(newState, httpStatus.OK, 'State created successfully'))
   } catch (error: any) {
-    return res.status(400).json({ error: error.message })
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).json(
+      createError(httpStatus.INTERNAL_SERVER_ERROR, error.message as string)
+    )
   }
 }
 
 export const updateState = async (req: Request, res: Response): Promise<Response> => {
-  const tenantId = res.locals.tenant_id
   const stateId = parseInt(req.params.id)
   const { description } = req.body
   try {
+    const stateExists = await prisma.state.findUnique({
+      where: {
+        id: stateId
+      }
+    })
+
+    if (!stateExists) {
+      return res.status(httpStatus.NOT_FOUND).json(
+        createError(httpStatus.NOT_FOUND, 'State not found')
+      )
+    }
+
     const updatedState = await prisma.state.update({
       where: {
-        tenant_id: tenantId,
         id: stateId
       },
       data: {
         description
       }
     })
-    return res.json(updatedState)
+    return res.status(httpStatus.OK).json(successResponse(updatedState, httpStatus.OK, 'State updated successfully'))
   } catch (error: any) {
-    return res.status(400).json({ error: error.message })
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).json(
+      createError(httpStatus.INTERNAL_SERVER_ERROR, error.message as string)
+    )
   }
 }
 
 export const deleteState = async (req: Request, res: Response): Promise<Response> => {
-  const tenantId = res.locals.tenant_id
   const stateId = parseInt(req.params.id)
   try {
-    await prisma.state.delete({
+    const stateExists = await prisma.state.findUnique({
       where: {
-        tenant_id: tenantId,
         id: stateId
       }
     })
-    return res.status(204).send()
+
+    if (!stateExists) {
+      return res.status(httpStatus.NOT_FOUND).json(
+        createError(httpStatus.NOT_FOUND, 'State not found')
+      )
+    }
+    await prisma.state.delete({
+      where: {
+        id: stateId
+      }
+    })
+    return res.status(httpStatus.OK).json(successResponse(null, httpStatus.OK, `State ${stateId} deleted successfully`))
   } catch (error: any) {
-    return res.status(500).json({ error: error.message })
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).json(
+      createError(httpStatus.INTERNAL_SERVER_ERROR, error.message as string)
+    )
   }
 }
